[Feat] 일정 상세 경로로 구현
일정 상세 경로로 구현
